SmartShoes_AI_Service/requirements.txt
fastapi
uvicorn[standard]
pandas
sentence-transformers
faiss-cpu
numpy
torch
transformers
requests
pydantic
rapidfuzz